+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Property `onViewStateChange` of all viewers accepts a callback for deck.gl view state changes.
 - Add a click handler to the overview that centers the detail view on the click position. The handler can be turned off by setting the `PictureInPictureViewer` property `clickCenter` to `false`.
 
 ### Changed
@@ -15,7 +16,7 @@
 
 ### Added
 
-- Export a `getDefaultInitialViewState` function for getting a deafult inital view state given a loader and desired size of view.
+- Export a `getDefaultInitialViewState` function for getting a default initial view state given a loader and desired size of view.
 - Support interleaved RGB OME-TIFF files.
 - Added a README file for Avivator at `avivator/README.md`.